Share vhost thread between mutiple devices
Vanilla vhost creates a worker thread and worker queue per virtio device.
This patch creates a worker thread and worker queue shared across multiple
virtio devices (VMs). The number of maximum virtio devices per worker can be
specified using a kernel module parameter.

Every time a new virtio device is created we search for a running worker
thread that is serving less than the maximum number of devices. If we find
such worker thread, we bind the new virtio device to this thread. Otherwise,
we create a new worker thread to serve the new virtio devices and more
devices that may be created in the future.

Currently, once a device is bound to a specific worker thread it can not be
migrated to other worker thread. In the future, we should improve the
mechanism to balance the devices across threads (migrate a device to a
different worker thread) and to create/destroy threads dynamically during
runtime based on the I/O activity.

Change-Id: I93691c84dd59ae16aa9774197a187e0df312f8b7
Signed-off-by: Abel Gordon <abelg@il.ibm.com>
Signed-off-by: Joel Nider <joeln@il.ibm.com>
Signed-off-by: Yossi Kuperman <yossiku@il.ibm.com>