Add herustics to improve I/O scheduling
This patch enhaces the round-robin mechanism with a set of heuristics to
decide when to leave a virtqueue and proceed to the next. The patch also
introduces a set of kernel module parameters to configure these heuristics.

The vhost generic layer exposes a new function that implements the heuristics.
The concrete code (e.g. vhost-net or vhost-blk) is required to call this function
to check when it should stop processing data.

The heuristics work as follows:

(1) We always leave a queue after doing a certain maximum
amount of work on it, even if it is not yet empty. This limitation is required
to avoids starvation.
(2) We may leave a queue earlier if we recognize that another non-empty queue
or the work queue is stuck and therefore likely to be latency-sensitive.
We call a queue stuck  if a certain time  has passed since it last received new
work. A latency-sensitive workload which waits for replies before sending further
requests will get "stuck" this sense, while a high-throughput workload which
continuously creates  new requests will not be found stuck.
(3) If we leave a queue erarlier without processing half of the maximum data, we
move the queue to the head of the round-robin list. We use this condition to
avoid degrading queues that were limited because other queues were stuck.
(4) Usually a bursty queues will add lot of items per burst. The time between
bursts may be bigger than the time we specified to detect stuck queues.
If a queue has more items than the specified in a module param, we never
consider the queue as stuck.  We use this condition to avoid detecting a
bursty queue as a stuck queue.
(5) We do not leave a queue before we did some minimum amount of work on it.
This technique improves cache efficiency and limits the number of queue switches
(important to improve throughput).

Change-Id: Ibeabed3fe4377a3fb4526ffdd721be2f90f62a51
Signed-off-by: Abel Gordon <abelg@il.ibm.com>
Signed-off-by: Joel Nider <joeln@il.ibm.com>
Signed-off-by: Yossi Kuperman <yossiku@il.ibm.com>