Major revisition of vhost controls
Everything is sitting under /proc/vhost/:
The structure is as follows:

/proc/vhost/worker/create:
Reading creates a new worker and returns the newly created worker id.

/proc/vhost/worker/remove:
Writing a worker id removes a worker with the same  id. The worker is
removed only if locked and have no queues assigned to it.

/proc/vhost/worker/default:
Writing a worker id set the worker with the same id as the designated
worker new devices are assigned to. Worker must be unlocked.

/proc/vhost/worker/<worker id>/locked:
Writing a bool sets worker as locked/unlocked, when a worker is locked
it cannot be assigned queues.
Reading returns 1 if device is locked, 0 otherwise.
Note: once a device has been locked it cannot be unlocked.

/proc/vhost/worker/<worker id>/cpu: (unimplemented)
Writing sets worker core affinity.
Reading returns the core affinity of the worker.

/proc/vhost/worker/<worker id>/exclusive: (unimplemented)
Writing a bool sets the worker as exclusive/non-exclusive in his core.
When a worker is exclusive no other thread is run in the same process.
Reading returns 1 if device is exclusive, 0 otherwise.

/proc/vhost/worker/<worker id>/dev_list:
Reading returns a '\n' separated list of ids of devices assigned to this worker.

/proc/vhost/worker/<device id>/worker:
Writing a worker id transfers a device from its current worker to the worker
with the writen id. The receiving worker cannot be a locked worker.
Reading returns the id of the worker the device is assigned to.

/proc/vhost/worker/<device id>/vq_list:
Reading returns a '\n' separated list of ids of queues contained in the
device.

/proc/vhost/vq/<vq id>/poll:
Writing starts/stops polling of virtual queue.
Reading returns the virtual queue current poll state.

/proc/vhost/vq/<vq id>/dev:
Reading returns the id of the device this queue is contained in.

Signed-off-by: Yossi Kuperman <yossiku@il.ibm.com>